Scheduler: Separate scheduler metrics module
