+++ core/src/banking_stage/transaction_scheduler/mod.rs
@@ -4,6 +4,7 @@ mod in_flight_tracker;
 pub(crate) mod prio_graph_scheduler;
 pub(crate) mod scheduler_controller;
 pub(crate) mod scheduler_error;
+mod scheduler_metrics;
 mod thread_aware_account_locks;
 mod transaction_id_generator;
 mod transaction_priority_id;
